Created a Wdgt with hist and text linked
Created a widget (HistoWithEditboxes) that contains an histogram (with
borders) and 2 lineEdit (that need to be changed to spinBoxes).

When we play with borders, the coordinate of the border is updated at
the same time in the corresponding lineEdit.

The lineEdits are read only at the moment.
Need to link them to the histograms.