Improved the picviewer
Internal objects (scaler and pixmapsize) are not reconstructed but
updated to improve performance.

Solved background not being updated (autoFillBackground = true)

Solved scaling issue. We must reset the pixmapsize instead of scaling it
again and again. Indeed when the display is very small the round on
pixels created an aspect ratio bug.

Created a custom 'update' function to improve code readability.
It calls 'super' update to make it happen.