Changed strategy for cv::Mat in pic viewer
Instead of subclassing OpenGLPicViewer to create a version that converts
cv::Mat into QPixmap, a dedicated converter class was created.
It enables to embeed the converter in the main viewer
(GLPicViewerWithCommands).

OpenGLPicViewer focuses on displaying QPixmap.
If more converters are needed they can be added in
GLPicViewerWithCommands.

Note: the converter is created only if needed.